+++ packages/zoltan2/src/util/Zoltan2_Util.hpp
@@ -89,7 +89,7 @@ template <typename Adapter, typename Extra>
   //
   const int *procList         = solution.getProcList();
 
-  int localSend = ((xtraInfo.size() == localNumIds) ? 1 : 0);
+  int localSend = ((size_t(xtraInfo.size()) == localNumIds) ? 1 : 0);
   int globalSend =0;
   reduceAll<int, int>(*comm, Teuchos::REDUCE_SUM, 1, &localSend, &globalSend);
 
@@ -104,7 +104,7 @@ template <typename Adapter, typename Extra>
       counts[partList[i]]++;
 
   Array<lno_t> offsets(numProcs+1, 0);
-  for (size_t i=1; i <= numProcs; i++){
+  for (int i=1; i <= numProcs; i++){
     offsets[i] = offsets[i-1] + counts[i-1];
   }
